src/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.10"
anyio==4.11.0 ; python_version >= "3.10"
boto3==1.40.74 ; python_version >= "3.10"
botocore==1.40.74 ; python_version >= "3.10"
cachetools==6.2.2 ; python_version >= "3.10"
certifi==2025.11.12 ; python_version >= "3.10"
charset-normalizer==3.4.4 ; python_version >= "3.10"
colorama==0.4.6 ; python_version >= "3.10" and platform_system == "Windows"
exceptiongroup==1.3.0 ; python_version == "3.10"
google-ai-generativelanguage==0.6.15 ; python_version >= "3.10"
google-api-core==2.25.2 ; python_version >= "3.14"
google-api-core==2.28.1 ; python_version >= "3.10" and python_version <= "3.13"
google-api-python-client==2.187.0 ; python_version >= "3.10"
google-auth-httplib2==0.2.1 ; python_version >= "3.10"
google-auth==2.43.0 ; python_version >= "3.10"
google-generativeai==0.8.5 ; python_version >= "3.10"
googleapis-common-protos==1.72.0 ; python_version >= "3.10"
grpcio-status==1.71.2 ; python_version >= "3.10"
grpcio==1.76.0 ; python_version >= "3.10"
h11==0.16.0 ; python_version >= "3.10"
httpcore==1.0.9 ; python_version >= "3.10"
httplib2==0.31.0 ; python_version >= "3.10"
httpx==0.28.1 ; python_version >= "3.10"
idna==3.11 ; python_version >= "3.10"
jmespath==1.0.1 ; python_version >= "3.10"
proto-plus==1.26.1 ; python_version >= "3.10"
protobuf==5.29.5 ; python_version >= "3.10"
pyasn1-modules==0.4.2 ; python_version >= "3.10"
pyasn1==0.6.1 ; python_version >= "3.10"
pydantic-core==2.41.5 ; python_version >= "3.10"
pydantic==2.12.4 ; python_version >= "3.10"
pyparsing==3.2.5 ; python_version >= "3.10"
python-dateutil==2.9.0.post0 ; python_version >= "3.10"
python-telegram-bot==22.5 ; python_version >= "3.10"
requests==2.32.5 ; python_version >= "3.10"
rsa==4.2 ; python_version >= "3.14"
rsa==4.9.1 ; python_version >= "3.10" and python_version <= "3.13"
s3transfer==0.14.0 ; python_version >= "3.10"
six==1.17.0 ; python_version >= "3.10"
sniffio==1.3.1 ; python_version >= "3.10"
tqdm==4.67.1 ; python_version >= "3.10"
typing-extensions==4.15.0 ; python_version >= "3.10"
typing-inspection==0.4.2 ; python_version >= "3.10"
uritemplate==4.2.0 ; python_version >= "3.10"
urllib3==2.5.0 ; python_version >= "3.10"
